+++ src/splat/util/options.py
@@ -364,6 +364,10 @@ def _parse_yaml(
         else:
             raise ValueError(f"Invalid endianness: {endianness}")
 
+    default_ld_bss_is_noload = True
+    if platform == "psx":
+        default_ld_bss_is_noload = False
+
     ret = SplatOpts(
         verbose=verbose,
         dump_symbols=p.parse_opt("dump_symbols", bool, False),
@@ -448,7 +452,7 @@ def _parse_yaml(
         ),
         ld_rom_start=p.parse_opt("ld_rom_start", int, 0),
         ld_fill_value=p.parse_optional_opt_with_default("ld_fill_value", int, 0),
-        ld_bss_is_noload=p.parse_opt("ld_bss_is_noload", bool, True),
+        ld_bss_is_noload=p.parse_opt("ld_bss_is_noload", bool, default_ld_bss_is_noload),
         ld_align_segment_vram_end=p.parse_opt("ld_align_segment_vram_end", bool, True),
         ld_align_section_vram_end=p.parse_opt("ld_align_section_vram_end", bool, True),
         ld_generate_symbol_per_data_segment=p.parse_opt(